Move test that is not supported in spatial
